Move parsing from gstttmlparse to gstttmlcommon to avoid multiple instances of static variables from timedText.
