+++ libs/gst/subtitle/parser/SubtitleParser.cpp
