+++ libs/gst/subtitle/parser/SubtitleParserCWrapper.cpp
