+++ libs/gst/subtitle/parser/SubtitleParserUtils.cpp
